+++ src/PreviewTag.ts
@@ -1,0 +1,5 @@
+export type PreviewTag =
+  | "ThumbnailImage"
+  | "PreviewImage"
+  | "JpgFromRaw"
+  | string